+++ packages/docs/site/src/routes/components/_examples/ui.js
@@ -671,4 +671,98 @@ export default formatExamples([
 		slug: 'ui-HighlightedText',
 		title: 'HighlightedText',
 	},
+	{
+		data: [{
+			key: `getText: () => 'Hello world!'`,
+			props: {
+				getText: () => 'Hello world!'
+			},
+			usage: `
+				<script>
+					import {CopyToClipboard} from '@svizzle/ui';
+				</script>
+
+				<CopyToClipboard getText={() => 'Hello world!'} />
+			`,
+		}, {
+			key: 'No `getText`',
+			props: {},
+			usage: `
+				<script>
+					import {CopyToClipboard} from '@svizzle/ui';
+				</script>
+
+				<CopyToClipboard />
+			`,
+		}, {
+			key: 'Theming (success)',
+			props: {
+				getText: () => 'Hello world!',
+				theme: {
+					background: 'antiquewhite',
+					color: 'brown',
+					failureColor: 'magenta',
+					outlineColor: 'blue',
+					outlineStyle: 'dotted',
+					outlineWidth: '2px',
+					successColor: 'palegreen',
+				}
+			},
+			usage: `
+				<script>
+					import {CopyToClipboard} from '@svizzle/ui';
+				</script>
+
+				<CopyToClipboard
+					getText={() => 'Hello world!'}
+					theme={{
+						background: 'antiquewhite',
+						color: 'brown',
+						failureColor: 'magenta',
+						outlineColor: 'blue',
+						outlineStyle: 'dotted',
+						outlineWidth: '2px',
+						successColor: 'palegreen',
+					}}
+				/>
+			`,
+		}, {
+			key: 'Theming (failure)',
+			props: {
+				theme: {
+					background: 'yellow',
+					color: 'brown',
+					failureColor: 'magenta',
+					outlineColor: 'blue',
+					outlineStyle: 'dotted',
+					outlineWidth: '2px',
+					successColor: 'palegreen',
+				}
+			},
+			usage: `
+				<script>
+					import {CopyToClipboard} from '@svizzle/ui';
+				</script>
+
+				<CopyToClipboard
+					theme={{
+						background: 'yellow',
+						color: 'brown',
+						failureColor: 'magenta',
+						outlineColor: 'blue',
+						outlineStyle: 'dotted',
+						outlineWidth: '2px',
+						successColor: 'palegreen',
+					}}
+				/>
+			`,
+		}],
+		doc: [
+			{tag: 'p', content: 'We need to pass a function that returns the text to be copied to the `getText` prop.'},
+		],
+		name: 'CopyToClipboard',
+		packageName: 'ui',
+		slug: 'ui-CopyToClipboard',
+		title: 'CopyToClipboard',
+	},
 ]);